+++ web/test/CredentialTemplates.test.ts
@@ -4,6 +4,7 @@ import {
     TrinsicService,
     CreateCredentialTemplateRequest,
     UpdateCredentialTemplateRequest,
+    UpdateStatusRequest,
     FieldType,
 } from "../src";
 // @ts-ignore
